feat(frames): update content in order to use frames from codesandbox
