+++ src/content/data/Icon/index.tsx
@@ -23,6 +23,20 @@ const icon = {
   description: "Icons used to communicate actions and decisions graphically",
   example: Icon,
   name: "Icon",
+  frame: (
+    <iframe
+      src="https://codesandbox.io/embed/pqwn7m?view=split&hidenavigation=0+%2B+preview&module=%2Fsrc%2FApp.tsx"
+      style={{
+        width: "100%",
+        height: "500px",
+        border: "0",
+        borderRadius: "4px",
+        overflow: "hidden",
+      }}
+      title="Avatar"
+      sandbox="allow-forms allow-modals allow-popups allow-presentation allow-same-origin allow-scripts"
+    />
+  ),
   props: {
     appearance: "primary",
     icon: MdAdd,
