+++ src/content/feedback/Assisted/index.tsx
@@ -57,6 +57,20 @@ const assisted = {
   description: "This component uses a filled primary icon for all applications",
   example: AssistedController,
   name: "Assisted",
+  frame: (
+    <iframe
+      src="https://codesandbox.io/embed/lqknjf?view=split&hidenavigation=0+%2B+preview&module=%2Fsrc%2FApp.tsx"
+      style={{
+        width: "100%",
+        height: "500px",
+        border: "0",
+        borderRadius: "4px",
+        overflow: "hidden",
+      }}
+      title="Avatar"
+      sandbox="allow-forms allow-modals allow-popups allow-presentation allow-same-origin allow-scripts"
+    />
+  ),
   props: {
     steps: stepsMock,
     currentStepId: 3,
