+++ src/content/feedback/SkeletonIcon/index.tsx
@@ -19,6 +19,20 @@ const skeletonIcon = {
   description: "This component uses a filled primary icon for all applications",
   example: SkeletonIcon,
   name: "SkeletonIcon",
+  frame: (
+    <iframe
+      src="https://codesandbox.io/embed/cpzntq?view=split&hidenavigation=0+%2B+preview&module=%2Fsrc%2FApp.tsx"
+      style={{
+        width: "100%",
+        height: "500px",
+        border: "0",
+        borderRadius: "4px",
+        overflow: "hidden",
+      }}
+      title="Avatar"
+      sandbox="allow-forms allow-modals allow-popups allow-presentation allow-same-origin allow-scripts"
+    />
+  ),
   installation: "npm install @inubekit/skeleton",
   props: {
     size: "24px",
