+++ src/content/feedback/SkeletonLine/index.tsx
@@ -19,6 +19,20 @@ const skeletonLine = {
   description: "This component uses a filled primary icon for all applications",
   example: SkeletonLine,
   name: "SkeletonLine",
+  frame: (
+    <iframe
+      src="https://codesandbox.io/embed/drk3jh?view=split&hidenavigation=0+%2B+preview&module=%2Fsrc%2FApp.tsx"
+      style={{
+        width: "100%",
+        height: "500px",
+        border: "0",
+        borderRadius: "4px",
+        overflow: "hidden",
+      }}
+      title="Avatar"
+      sandbox="allow-forms allow-modals allow-popups allow-presentation allow-same-origin allow-scripts"
+    />
+  ),
   installation: "npm install @inubekit/skeleton",
   props: {
     width: "100%",
