+++ src/content/inputs/Label/index.tsx
@@ -22,6 +22,20 @@ const label = {
   description: "This component uses a filled primary icon for all applications",
   example: LabelController,
   name: "Label",
+  frame: (
+    <iframe
+      src="https://codesandbox.io/embed/qcq5sm?view=split&hidenavigation=0+%2B+preview&module=%2Fsrc%2FApp.tsx"
+      style={{
+        width: "100%",
+        height: "500px",
+        border: "0",
+        borderRadius: "4px",
+        overflow: "hidden",
+      }}
+      title="Avatar"
+      sandbox="allow-forms allow-modals allow-popups allow-presentation allow-same-origin allow-scripts"
+    />
+  ),
   props: {
     htmlFor: "id",
     children: "Label Text",
