+++ src/content/inputs/Toggle/index.tsx
@@ -20,6 +20,20 @@ const toggle = {
   description: "This component uses a filled primary icon for all applications",
   example: ToggleController,
   name: "Toggle",
+  frame: (
+    <iframe
+      src="https://codesandbox.io/embed/s8j7wd?view=split&hidenavigation=0+%2B+preview&module=%2Fsrc%2FApp.tsx"
+      style={{
+        width: "100%",
+        height: "500px",
+        border: "0",
+        borderRadius: "4px",
+        overflow: "hidden",
+      }}
+      title="Avatar"
+      sandbox="allow-forms allow-modals allow-popups allow-presentation allow-same-origin allow-scripts"
+    />
+  ),
   props: {
     id: "id",
     disabled: false,
