+++ src/content/layout/Divider/index.tsx
@@ -20,6 +20,20 @@ const divider = {
     "Element that helps separate pieces of content into groups, sections, options or parts, in a clear and defined way.",
   example: DividerController,
   name: "Divider",
+  frame: (
+    <iframe
+      src="https://codesandbox.io/embed/s9hnj9?view=split&hidenavigation=0+%2B+preview&module=%2Fsrc%2FApp.tsx"
+      style={{
+        width: "100%",
+        height: "500px",
+        border: "0",
+        borderRadius: "4px",
+        overflow: "hidden",
+      }}
+      title="Avatar"
+      sandbox="allow-forms allow-modals allow-popups allow-presentation allow-same-origin allow-scripts"
+    />
+  ),
   props: {
     dashed: false,
   },
