+++ src/content/layout/Stack/index.tsx
@@ -4,6 +4,20 @@ const stack = {
   description: "This component uses a filled primary icon for all applications",
   example: StackController,
   name: "Stack",
+  frame: (
+    <iframe
+      src="https://codesandbox.io/embed/cylxks?view=split&hidenavigation=0+%2B+preview&module=%2Fsrc%2FApp.tsx"
+      style={{
+        width: "100%",
+        height: "500px",
+        border: "0",
+        borderRadius: "4px",
+        overflow: "hidden",
+      }}
+      title="Avatar"
+      sandbox="allow-forms allow-modals allow-popups allow-presentation allow-same-origin allow-scripts"
+    />
+  ),
   props: {
     children: [...Array(6 + 1).keys()].slice(1),
     gap: "10px",
