+++ src/content/navigation/FullscreenNav/index.tsx
@@ -1,22 +1,19 @@
-import { Header } from "@inubekit/header";
-
+import { FullscreenNav } from "@inubekit/fullscreennav";
 import {
-  MdKey,
+  MdVpnKey,
   MdMeetingRoom,
-  MdPhone,
+  MdContacts,
   MdStarBorder,
   MdBadge,
   MdAccountBalanceWallet,
   MdAccountBalance,
-  MdAndroid,
 } from "react-icons/md";
-import { Logo } from "./logo";
 import { inube } from "@inubekit/foundations";
 import { buildTokenDescriptions } from "../../tokens/buildTokenDescriptions";
 
-const headerTokensConfig = {
+const fullscreenNavTokensConfig = {
   businessUnit: "inube",
-  component: "header",
+  component: "fullscreenNav",
   include: [
     { id: "businessUnit", order: 1 },
     { id: "component", order: 2 },
@@ -26,22 +23,36 @@ const headerTokensConfig = {
   ],
 };
 
-const header = {
+const fullscreenNav = {
   description: "This component uses a filled primary icon for all applications",
-  example: Header,
-  name: "Header",
+  example: FullscreenNav,
+  name: "FullscreenNav",
+  frame: (
+    <iframe
+      src="https://codesandbox.io/embed/7swgmz?view=split&hidenavigation=0+%2B+preview&module=%2Fsrc%2FApp.tsx"
+      style={{
+        width: "100%",
+        height: "500px",
+        border: "0",
+        borderRadius: "4px",
+        overflow: "hidden",
+      }}
+      title="Avatar"
+      sandbox="allow-forms allow-modals allow-popups allow-presentation allow-same-origin allow-scripts"
+    />
+  ),
   props: {
     portalId: "portals",
     navigation: {
       title: "MENU",
       sections: {
-        administrate: {
-          name: "Administrate",
+        administrative: {
+          name: "Administrative",
           links: {
             privileges: {
               id: "privileges",
               label: "Privileges",
-              icon: MdKey,
+              icon: MdVpnKey,
               path: "/privileges",
             },
             accounting: {
@@ -53,7 +64,7 @@ const header = {
             contacts: {
               id: "contacts",
               label: "Contacts",
-              icon: MdPhone,
+              icon: MdContacts,
               path: "/contacts",
             },
             crm: {
@@ -95,18 +106,8 @@ const header = {
         },
       },
     },
-    logoURL: Logo,
-    userName: "Leonardo Garzón",
-    client: "Sistemas Enlínea S.A",
-    links: [
-      {
-        label: "Actualizar datos",
-        path: "/update-data-assisted",
-        icon: MdAndroid,
-      },
-    ],
-    showLinks: false,
-    showUser: false,
+    logoutPath: "/logout",
+    logoutTitle: "logout",
   },
   propTypes: {
     portalId: {
@@ -114,31 +115,21 @@ const header = {
     },
     navigation: {
       description:
-        "The primary object that will organize and store the requisite paths for the correct operation of the Nav component is forthcoming and is required",
+        "shall be designed to accept an array of objects with a predetermined structure, as specified below: Each object shall contain the following attributes",
     },
-    logoURL: {
-      description:
-        "prop accepts a component to be used as the logo in the header. This component can be an image, an icon, stylized text or any other visual element that represents the brand identity.",
-    },
-    userName: {
-      description: "shall be the displayed username",
+    logoutPath: {
+      description: "path to the logout page",
       type: "input",
     },
-    client: {
-      description: "shall be the displayed business Unit",
+    logoutTitle: {
+      description: "title of the logout page",
       type: "input",
-    },
-    links: {
-      description: "shall be the links that it'll be shown in the header",
-    },
-    showLinks: {
-      description: "it determines if the links will be shown in the header",
-    },
-    showUser: {
-      description: "it determines if the user will be shown in the header",
     },
   },
-  tokens: buildTokenDescriptions(inube.header, headerTokensConfig),
+  tokens: buildTokenDescriptions(
+    inube.fullscreenNav,
+    fullscreenNavTokensConfig,
+  ),
 };
 
-export { header };
+export { fullscreenNav };
