+++ src/content/navigation/Link/index.tsx
@@ -20,6 +20,20 @@ const link = {
   description: "This component uses a filled primary icon for all applications",
   example: LinkController,
   name: "Link",
+  frame: (
+    <iframe
+      src="https://codesandbox.io/embed/f54pvj?view=split&hidenavigation=0+%2B+preview&module=%2Fsrc%2FApp.tsx"
+      style={{
+        width: "100%",
+        height: "500px",
+        border: "0",
+        borderRadius: "4px",
+        overflow: "hidden",
+      }}
+      title="Avatar"
+      sandbox="allow-forms allow-modals allow-popups allow-presentation allow-same-origin allow-scripts"
+    />
+  ),
   props: {
     children: "Link",
     size: "large",
