+++ src/content/navigation/Nav/index.tsx
@@ -28,6 +28,20 @@ const nav = {
   description: "This component uses a filled primary icon for all applications",
   example: NavController,
   name: "Nav",
+  frame: (
+    <iframe
+      src="https://codesandbox.io/embed/2hd7vv?view=split&hidenavigation=0+%2B+preview&module=%2Fsrc%2FApp.tsx"
+      style={{
+        width: "100%",
+        height: "500px",
+        border: "0",
+        borderRadius: "4px",
+        overflow: "hidden",
+      }}
+      title="Avatar"
+      sandbox="allow-forms allow-modals allow-popups allow-presentation allow-same-origin allow-scripts"
+    />
+  ),
   props: {
     navigation: {
       title: "MENU",
