+++ src/content/navigation/Tabs/index.tsx
@@ -20,6 +20,20 @@ const tabs = {
   description: "This component uses a filled primary icon for all applications",
   example: TabsController,
   name: "Tabs",
+  frame: (
+    <iframe
+      src="https://codesandbox.io/embed/d2xvgv?view=split&hidenavigation=0+%2B+preview&module=%2Fsrc%2FApp.tsx"
+      style={{
+        width: "100%",
+        height: "500px",
+        border: "0",
+        borderRadius: "4px",
+        overflow: "hidden",
+      }}
+      title="Avatar"
+      sandbox="allow-forms allow-modals allow-popups allow-presentation allow-same-origin allow-scripts"
+    />
+  ),
   props: {
     tabs: [
       {
